[camera_avfoundation] Pigeon swift migration - part 1 (#10939)
Prepares camera package for pigeon Swift migration as part of https://github.com/flutter/flutter/issues/119109

Current implementation of `FormatUtils.selectBestFormat` overrides `MediaSettings.framesPerSecond`. With swift pigeon contracts `MediaSettings` is non mutable so that will no longer be possible.

## Pre-Review Checklist

**Note**: The Flutter team is currently trialing the use of [Gemini Code Assist for GitHub](https://developers.google.com/gemini-code-assist/docs/review-github-code). Comments from the `gemini-code-assist` bot should not be taken as authoritative feedback from the Flutter team. If you find its comments useful you can update your code accordingly, but if you are unsure or disagree with the feedback, please feel free to wait for a Flutter team member's review for guidance on which automated comments should be addressed.

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.

+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraSettingsTests.swift
@@ -86,8 +86,8 @@ private final class TestMediaSettingsAVWrapper: FLTCamMediaSettingsAVWrapper {
     if let compressionProperties = outputSettings?[AVVideoCompressionPropertiesKey]
       as? [String: Any],
       let bitrate = compressionProperties[AVVideoAverageBitRateKey] as? Int,
-      let frameRate = compressionProperties[AVVideoExpectedSourceFrameRateKey] as? Int,
-      bitrate == testVideoBitrate, frameRate == testFramesPerSecond
+      let frameRate = compressionProperties[AVVideoExpectedSourceFrameRateKey] as? Double,
+      bitrate == testVideoBitrate, frameRate == Double(testFramesPerSecond)
     {
       videoSettingsExpectation.fulfill()
     }
